Docs: Update atlas_artifact resource to use data
A while back `atlas_artifact` was switched from being a `resource` to a `data` provider. When you use the examples suggested in the Terraform Enterprise docs, the Terraform cli shows a deprecation warning and provides an old url to the new data provider docs.

There are some complimentary doc updates in the Terraform Enterprise/Atlas repo.